requirements.txt
asgiref==3.6.0
Django==4.2.1
django-autoslug==1.9.8
django-ckeditor-5==0.2.3
django-colorfield==0.8.0
django-cors-headers==3.13.0
django-filter==22.1
django-jazzmin==2.6.0
django-translated-fields==0.12.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.4.0
PyJWT==2.6.0
python-decouple==3.7
python-dotenv==0.21.1
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
Unidecode==1.3.6
whitenoise==6.3.0
